requirements.txt
Flask==3.0.3
gunicorn==22.0.0
PyMuPDF==1.24.5
python-dotenv==1.0.1
google-generativeai==0.7.0
rouge-score